Refactored FifsUtility.sol to Utility.sol
* Split interface IUtility.sol into IUtilityBase and IUtility
* IUtilityBase is mostly IUtility but without settle function
* IUtiilty extends from IUtilityBase
* Remove interface modifier to make inheritance possible
* Adapted this pattern from https://github.com/OpenZeppelin/openzeppelin-solidity/blob/master/contracts/token/ERC721/IERC721.sol
